move validate functionality into lib folder
