+++ api/src/license/types/index.ts
@@ -1,0 +1,1 @@
+export * from './validate.js';
