+++ api/src/license/types/validate.ts
